Fix pointer truncation when buildings on x64 Windows
The fix is (long) → (size_t) everywhere, matching what gperf already generates in aliases.h. On Win64 LLP64, long is 32 bits but size_t is 64 bits (pointer-width).

Fixes several warnings like

```
lib\canonical.h(110,8): warning C4311: 'type cast': pointer truncation from 'char (*)[6]' to 'long'
```

+++ lib/canonical_local_sysaix.h
@@ -1,2 +1,2 @@
-  (int)(long)&((struct stringpool_t *)0)->stringpool_str15,
-  (int)(long)&((struct stringpool_t *)0)->stringpool_str484,
+  (int)(size_t)&((struct stringpool_t *)0)->stringpool_str15,
+  (int)(size_t)&((struct stringpool_t *)0)->stringpool_str484,
