+++ lib/canonical_local_syssolaris.h
@@ -1,2 +1,2 @@
-  (int)(long)&((struct stringpool_t *)0)->stringpool_str112,
-  (int)(long)&((struct stringpool_t *)0)->stringpool_str491,
+  (int)(size_t)&((struct stringpool_t *)0)->stringpool_str112,
+  (int)(size_t)&((struct stringpool_t *)0)->stringpool_str491,
